requirements.txt
# Core PostgreSQL
psycopg2-binary==2.9.9
psycopg[binary]==3.1.12

# LangChain 관련 (기존 버전 유지)
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4

# OpenAI
openai==1.45.0

# 데이터 처리
pandas==2.2.2
numpy==1.26.4

# FastAPI (기존 버전 유지)
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# 추가 FastAPI 관련 패키지
python-multipart==0.0.6
pydantic-settings==2.5.0

# 유틸리티
python-dotenv==1.0.1
sqlalchemy==2.0.35

# 로깅 및 모니터링
structlog==24.4.0


###개발용 의존성(테스트 등)
# 테스트
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0

# 코드 품질
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2

# 개발 도구
pre-commit==3.8.0

## 슬랙
slack_sdk